+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes for Lovely Gradle Plugin
 
+## Unreleased
+
+- allow to specify build platforms for local docker builds.
+
 ## 2022-08-29 / 1.8.1
 
 - fix: Prefer the usage of the custom docker-container builder for local builds